# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WhyNoIPv6 API
  description: >
    The public read API of whynoipv6.com (07-api.md). Versionless root,
    snake_case wire, `{items,page,meta}` / `{points,meta}` envelopes, keyset
    cursor pagination, RFC 9457 problem+json errors. The only write path is
    the anonymous async live check (`POST /check`). Data is licensed
    CC-BY-NC-4.0.
  version: "1.0.0"
  contact:
    url: https://whynoipv6.com
  license:
    name: CC-BY-NC-4.0
    url: https://creativecommons.org/licenses/by-nc/4.0/
servers:
  - url: https://api.whynoipv6.com
# The filter composition every tier/scoped list shares with /domains
# (07 §4.4). YAML anchor — expanded per path at parse time.
...
paths:
...
  /countries/{code}/stats:
    get:
      operationId: getCountryStats
      tags: [stats]
      summary: Country adoption time series
      parameters:
        - $ref: "#/components/parameters/countryCode"
        - $ref: "#/components/parameters/statsFrom"
        - $ref: "#/components/parameters/statsTo"
        - $ref: "#/components/parameters/interval"
      responses:
        "200":
          description: Confirmed-state snapshots.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CountryStatsSeries" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /asns/{number}/stats:
    get:
      operationId: getASNStats
      tags: [stats]
      summary: Network adoption time series (canonical count_v6/count_total wire names)
      parameters:
        - name: number
          in: path
          required: true
          description: AS number; an optional leading `AS` prefix is accepted.
          schema: { type: string }
        - $ref: "#/components/parameters/statsFrom"
        - $ref: "#/components/parameters/statsTo"
        - $ref: "#/components/parameters/interval"
      responses:
        "200":
          description: Confirmed-state snapshots.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ASNStatsSeries" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /campaigns/{uuid}/stats:
    get:
      operationId: getCampaignStats
      tags: [stats]
      summary: Campaign adoption time series (incl. v6_ready)
      parameters:
        - $ref: "#/components/parameters/campaignUUID"
        - $ref: "#/components/parameters/statsFrom"
        - $ref: "#/components/parameters/statsTo"
        - $ref: "#/components/parameters/interval"
      responses:
        "200":
          description: Confirmed-state snapshots; disabled campaigns are 404.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CampaignStatsSeries" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [stats]
      summary: The headline adoption dashboard series
      parameters:
        - $ref: "#/components/parameters/statsFrom"
        - $ref: "#/components/parameters/statsTo"
        - $ref: "#/components/parameters/interval"
      responses:
        "200":
          description: "`meta.source` is always `confirmed_state`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GlobalStatsSeries" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
...
components:
  parameters:
...
    countryCode:
      name: code
      in: path
      required: true
      description: ISO 3166-1 alpha-2, case-insensitive; `UN` is the sentinel.
      schema: { type: string, minLength: 2, maxLength: 2 }
...
    campaignUUID:
      name: uuid
      in: path
      required: true
      description: The raw campaign UUID.
      schema: { type: string, format: uuid }
...
    statsFrom:
      name: from
      in: query
      description: Window start (YYYY-MM-DD); default `to − 90d`.
      schema: { type: string, format: date }
    statsTo:
      name: to
      in: query
      description: Window end (YYYY-MM-DD); default today UTC.
      schema: { type: string, format: date }
    interval:
      name: interval
      in: query
      description: "`weekly` keeps the latest snapshot per ISO week — a sample, never an average."
      schema: { type: string, enum: [daily, weekly], default: daily }
  responses:
...
    NotFound:
      description: Unknown or uncanonicalizable resource key.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InvalidParameter:
      description: Malformed cursor, bad format, malformed badge host, bad window.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Meta:
      type: object
      required: [as_of, generation, license]
      properties:
        as_of: { type: string, format: date-time }
        generation: { type: integer, description: "YYYYMMDD of the newest stats day", example: 20260707 }
        count: { type: integer, format: int64, description: Exact — bounded curated sets only. }
        count_estimate: { type: integer, format: int64, description: Approximate — everything else. }
        license: { type: string, example: CC-BY-NC-4.0 }
        source: { type: string, description: "`confirmed_state` on every time series." }
...
    Problem:
      type: object
      description: RFC 9457. `status` always equals the HTTP status line.
      required: [type, title, status]
      properties:
        type: { type: string, example: "https://whynoipv6.com/problems/not-found" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    GlobalStatsPoint:
      type: object
      required: [day]
      properties:
        day: { type: string, format: date }
        domains: { type: integer, nullable: true }
        heroes: { type: integer, nullable: true }
        partial: { type: integer, nullable: true }
        sinners: { type: integer, nullable: true }
        inactive: { type: integer, nullable: true }
        unknown: { type: integer, nullable: true }
        saints: { type: integer, nullable: true }
        disabled: { type: integer, nullable: true }
        base_supported: { type: integer, nullable: true }
        www_supported: { type: integer, nullable: true }
        ns_supported: { type: integer, nullable: true }
        mx_supported: { type: integer, nullable: true }
        conn_supported: { type: integer, nullable: true }
        resources_supported: { type: integer, nullable: true }
        top_heroes: { type: integer, nullable: true }
        top_nameserver: { type: integer, nullable: true }
    GlobalStatsSeries:
      type: object
      required: [points, meta]
      properties:
        points:
          type: array
          items: { $ref: "#/components/schemas/GlobalStatsPoint" }
        meta: { $ref: "#/components/schemas/Meta" }
    CountryStatsPoint:
      type: object
      required: [day]
      properties:
        day: { type: string, format: date }
        domains: { type: integer, nullable: true }
        sinners: { type: integer, nullable: true }
        partial: { type: integer, nullable: true }
        heroes: { type: integer, nullable: true }
        base_supported: { type: integer, nullable: true }
        conn_supported: { type: integer, nullable: true }
    CountryStatsSeries:
      type: object
      required: [points, meta]
      properties:
        points:
          type: array
          items: { $ref: "#/components/schemas/CountryStatsPoint" }
        meta: { $ref: "#/components/schemas/Meta" }
    CampaignStatsPoint:
      type: object
      required: [day]
      properties:
        day: { type: string, format: date }
        domains: { type: integer, nullable: true }
        v6_ready: { type: integer, nullable: true }
        sinners: { type: integer, nullable: true }
        partial: { type: integer, nullable: true }
        heroes: { type: integer, nullable: true }
        base_supported: { type: integer, nullable: true }
        www_supported: { type: integer, nullable: true }
        ns_supported: { type: integer, nullable: true }
        mx_supported: { type: integer, nullable: true }
        conn_supported: { type: integer, nullable: true }
    CampaignStatsSeries:
      type: object
      required: [points, meta]
      properties:
        points:
          type: array
          items: { $ref: "#/components/schemas/CampaignStatsPoint" }
        meta: { $ref: "#/components/schemas/Meta" }
    ASNStatsPoint:
      type: object
      description: "`v6_domains`/`domains` mapped onto the canonical count_v6/count_total wire names."
      required: [day]
      properties:
        day: { type: string, format: date }
        count_total: { type: integer, nullable: true }
        count_v6: { type: integer, nullable: true }
        sinners: { type: integer, nullable: true }
        heroes: { type: integer, nullable: true }
    ASNStatsSeries:
      type: object
      required: [points, meta]
      properties:
        points:
          type: array
          items: { $ref: "#/components/schemas/ASNStatsPoint" }
        meta: { $ref: "#/components/schemas/Meta" }
...